Change file monitor
Replace inja single include
