+++ src/config/config_modal.hpp
@@ -78,4 +78,3 @@ struct WebhookConfigModal
         return config;
     }
 };
-
